+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.3.3] - 2020-05-17
+
+- Improve the definition types of the method getHarmony
+- Fix a bug in the getHarmony method that was returning RGBObject colors when a HEXObject color was provided
+
 ## [1.3.2] - 2020-05-16
 
 - Fix type definitions bugs